2D Player Grid v2.1
Huge Change. I changed it from vectors to a 2d array. Works MUCH better vectors and way easier to code with. I also changed the way it changed your location on the graph when you press a button, since the previous method with vectors wouldn't allow barriers in the future.

Also, this is v2.1 because I forgot to add v2.0. After I made the change from vectors to a 2d array i also added messages to let you know how to exit the game and if you're at a border.