2D Player Grid v2.2
I added 2 constants; one for the max x size, and one for the max y size. This makes it much easier to change the size of the grid, so I increased the size from 10x10 to 16x16.

I also reduced the amount of arrays I declare in the beginning because if a unit is a 0 by default, then I only need to declare the first array since that is where the player's initial location is. I don't know how to explain correctly it but the point was to clean up the code a bit, however, keeping all 16 arrays in might give other people a better visual. I'll think about it.